@@ -4,27 +4,28 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.2
-  lint: ^1.5.3
-  firebase_core: ^1.3.0
-  firebase_auth: ^1.4.1
-  firebase_analytics: ^8.1.2
-  cloud_firestore: ^2.2.2
-  country_code_picker: ^2.0.2
+  lint: ^2.1.2
+  firebase_core: ^2.16.0
+  firebase_auth: ^4.10.0
+  firebase_analytics: ^10.5.0
+  cloud_firestore: ^4.9.2
+  country_code_picker: ^3.0.0
   google_maps_flutter: ^2.0.8
-  percent_indicator: ^3.0.1
+  percent_indicator: ^4.2.3
   flutter_google_places: ^0.3.0
   provider: ^6.0.1
-  geolocator: ^7.6.2
+  geolocator: ^10.0.1
   http: ^0.13.3
   flutter_spinkit: ^5.1.0
   geocoding: ^2.0.1
-  connectivity_plus: ^2.0.2
+  connectivity_plus: ^4.0.2
+  intl: ^0.18.1
 
 dev_dependencies:
   flutter_test:
@@ -37,5 +38,4 @@ flutter:
   assets:
     - assets/
     - assets/images/
-    - assets/icons/
  
