@@ -3,25 +3,25 @@ import 'package:ukulimabora/Shared/Common/constants.dart';
 
 class UkulimaBoraTextFormField extends StatelessWidget {
   const UkulimaBoraTextFormField(
-      {@required this.fillcolor,
-      @required this.controller,
-      @required this.hinttext,
-      @required this.keyboardtype,
-      @required this.validator,
-      @required this.onsaved,
-      @required this.onchanged,
-      this.maximumLength,
-      this.obscuretext,
-      this.style,
-      this.hintStyle,
-      this.icon});
+      {required this.fillcolor,
+      required this.controller,
+      required this.hinttext,
+      required this.keyboardtype,
+      required this.validator,
+      required this.onsaved,
+      required this.onchanged,
+      required this.maximumLength,
+      required this.obscuretext,
+      required this.style,
+      required this.hintStyle,
+      required this.icon,});
 
   final TextEditingController controller;
   final String hinttext;
   final Color fillcolor;
   final TextInputType keyboardtype;
-  final String Function(String) validator;
-  final Function(String) onsaved;
+  final String? Function(String?)? validator;
+  final Function(String?)? onsaved;
   final Function(String) onchanged;
   final int maximumLength;
   final bool obscuretext;
@@ -54,28 +54,28 @@ class UkulimaBoraTextFormField extends StatelessWidget {
               ),
               borderSide: BorderSide(
                 color: UkulimaBoraCommonColors.appBlueColor,
-              )),
+              ),),
           errorBorder: OutlineInputBorder(
               borderRadius: const BorderRadius.all(
                 Radius.circular(70.0),
               ),
               borderSide: BorderSide(
                 color: UkulimaBoraCommonColors.appRedColor,
-              )),
+              ),),
           enabledBorder: OutlineInputBorder(
               borderRadius: const BorderRadius.all(
                 Radius.circular(70.0),
               ),
               borderSide: BorderSide(
                 color: UkulimaBoraCommonColors.appGreenColor,
-              )),
+              ),),
           focusedBorder: OutlineInputBorder(
               borderRadius: const BorderRadius.all(
                 Radius.circular(70.0),
               ),
               borderSide: BorderSide(
                 color: UkulimaBoraCommonColors.appBlueColor,
-              ))),
+              ),),),
     );
   }
 }
