@@ -3,8 +3,8 @@ import 'package:ukulimabora/Shared/Common/constants.dart';
 
 class OnboardingButton extends StatefulWidget {
   const OnboardingButton({
-    @required this.buttonText,
-    @required this.buttonAction,
+    required this.buttonText,
+    required this.buttonAction,
   });
 
   final String buttonText;
@@ -23,14 +23,14 @@ class _OnboardingButtonState extends State<OnboardingButton> {
       width: MediaQuery.of(context).size.width,
       decoration: BoxDecoration(
           color: UkulimaBoraCommonColors.appGreenColor,
-          borderRadius: BorderRadius.circular(70)),
+          borderRadius: BorderRadius.circular(70),),
       child: TextButton(
           onPressed: widget.buttonAction,
           child: Text(
             widget.buttonText,
             style: TextStyle(
-                color: UkulimaBoraCommonColors.appWhiteColor, fontSize: 18),
-          )),
+                color: UkulimaBoraCommonColors.appWhiteColor, fontSize: 18,),
+          ),),
     );
   }
 }
