@@ -1,25 +1,25 @@
 import 'package:firebase_auth/firebase_auth.dart';
 
 class AuthenticationService {
-  AuthenticationService({this.firebaseAuth});
+ AuthenticationService({required this.firebaseAuth});
   final FirebaseAuth firebaseAuth;
 
-  Stream<User> get authStateChanges => firebaseAuth.authStateChanges();
+  Stream<User?> get authStateChanges => firebaseAuth.authStateChanges();
 
-  Future<String> signIn({String email, String password}) async {
+  Future<String?> signIn({required String email, required String password}) async{
     try {
       await firebaseAuth.signInWithEmailAndPassword(
-          email: email, password: password);
+          email: email, password: password,);
       return 'Signed In';
     } on FirebaseAuthException catch (e) {
       return e.message;
     }
   }
 
-  Future<String> signUp({String email, String password}) async {
+  Future<String?> signUp({required String email, required String password}) async {
     try {
       await firebaseAuth.createUserWithEmailAndPassword(
-          email: email, password: password);
+          email: email, password: password,);
       return 'Signed Up';
     } on FirebaseAuthException catch (e) {
       return e.message;
