@@ -3,6 +3,6 @@ import 'package:geolocator/geolocator.dart';
 class GeolocatorService {
   Future<Position> getCurrentLocation() async {
     return Geolocator.getCurrentPosition(
-        desiredAccuracy: LocationAccuracy.high);
+        desiredAccuracy: LocationAccuracy.high,);
   }
 }
