@@ -56,7 +56,7 @@ class _LandPrepSuggestionState extends State<LandPrepSuggestion> {
 
                               final DateTime date =
                                   DateTime.fromMillisecondsSinceEpoch(
-                                      timeStamp * 1000);
+                                      timeStamp * 1000,);
 
                               final String formatedDate =
                                   DateFormat('EEEE MMM-dd').format(date);
@@ -88,7 +88,7 @@ class _LandPrepSuggestionState extends State<LandPrepSuggestion> {
                                 );
                               }
                               return Container();
-                            }),
+                            },),
                       ),
                     )
                   ],
@@ -101,7 +101,7 @@ class _LandPrepSuggestionState extends State<LandPrepSuggestion> {
                 );
               }
               return UkulimaBoraLoadingIndicator();
-            }),
+            },),
       ),
     );
   }
